fix: SonarCloud code duplication Quality Gate fail
